refactor: jgrapht에서 Graph 라이브러리 주입받도록 변경
