+++ src/main/java/wooteco/subway/utils/ShortestPathEdge.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.utils;
 
 import org.jgrapht.graph.DefaultWeightedEdge;
